split command-line parsers & handlers into multiple organized modules
