+++ src/Command/Key/Public.hs
@@ -1,0 +1,51 @@
+{-# LANGUAGE NamedFieldPuns #-}
+{-# LANGUAGE OverloadedStrings #-}
+{-# LANGUAGE QuasiQuotes #-}
+
+{-# OPTIONS_HADDOCK hide #-}
+
+module Command.Key.Public
+    ( Cmd (..)
+    , mod
+    , run
+    ) where
+
+import Prelude hiding
+    ( mod )
+
+import Cardano.Address.Derivation
+    ( toXPub, xpubToBytes )
+import Codec.Binary.Bech32.TH
+    ( humanReadablePart )
+import Options.Applicative
+    ( CommandFields, Mod, command, footerDoc, helper, info, progDesc )
+import Options.Applicative.Encoding
+    ( Encoding, encodingOpt )
+import Options.Applicative.Help.Pretty
+    ( string )
+import System.IO
+    ( stdin, stdout )
+import System.IO.Extra
+    ( hGetXPrv, hPutBytes )
+
+
+newtype Cmd = Public
+    { encoding :: Encoding
+    } deriving (Show)
+
+mod :: (Cmd -> parent) -> Mod CommandFields parent
+mod liftCmd = command "public" $
+    info (helper <*> fmap liftCmd parser) $ mempty
+        <> progDesc "Get the public counterpart of a private key."
+        <> footerDoc (Just $ string $ mconcat
+            [ "The private key is read from stdin."
+            ])
+  where
+    parser = Public
+        <$> encodingOpt [humanReadablePart|xpub|]
+
+run :: Cmd -> IO ()
+run Public{encoding} = do
+    xprv <- hGetXPrv stdin
+    let xpub = toXPub xprv
+    hPutBytes stdout (xpubToBytes xpub) encoding
